% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_next.R
\name{antipodal_power}
\alias{antipodal_power}
\title{Dioptric power at the antipod of K-max}
\usage{
antipodal_power(exam_curvature, exam_k_max, interp = F)
}
\arguments{
\item{exam_k_max}{A data frame containing the \link{join_fields} and the \code{x}, \code{y},
and \code{measurement} of K-max.}

\item{interp}{A Boolean. \code{TRUE} to interpolate measurements. See details.}

\item{exam_record}{A data frame with the same structure as \link{sample_curvature},
containing one row for each curvature \code{measurement}.}
}
\value{
A one-row data frame containing the \link{join_fields} and the following
columns:

\describe{
\item{k_max}{Dioptric power of K-max.}
\item{opposite_power}{Dioptric power of the antipod.}
\item{x}{The \emph{x}-axis position of the measurement. See details.}
\item{y}{The \emph{y}-axis position of the measurement.}
\item{opposite_x}{The \emph{x}-axis position of the antipod. See details.}
\item{opposite_y}{The \emph{y}-axis position of the antipod.}
\item{euch_dist}{Euchlidean distance of the measurement from the antipod.}
}
}
\description{
Returns the dioptric power at the antipod of K-max.
}
\details{
The antipod of K-max is directly opposite K-max on the curvature map.

The \link{join_fields} in the returned data frame come from the
\code{exam_k_max}.
}
\seealso{
Other K-next: 
\code{\link{find_peaks}()},
\code{\link{k_next}()},
\code{\link{k_next_old}()}
}
\concept{K-next}
