% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematics.R
\name{array_position}
\alias{array_position}
\title{Returns the index/position of a 2d matrix for a given row and column}
\usage{
array_position(
  matrix,
  row_index = NULL,
  col_index = NULL,
  row_label = NULL,
  col_label = NULL
)
}
\arguments{
\item{matrix}{A two-dimensional array.}

\item{row_index}{The row number.}

\item{col_index}{The column number.}

\item{row_label}{A character value matching one of the row names of \code{matrix}.}

\item{col_label}{A character value matching one of the column names of \code{matrix}.}
}
\value{
Returns a numeric index (of the matrix).
}
\description{
This function does the inverse of \code{arrayInd(index, dim(array))}.
}
\details{
Requires both a row and column, which can be provided as a \code{row_index}
and \code{col_index}, as a \code{row_label} and \code{col_label}, or some
combination of an index and a label.
}
\examples{
# find the index of row 2, column 1
matrix(seq(1:9), nrow = 3, ncol = 3, byrow = T) |>
  array_position(row_index = 2, col_index = 1)

# find the row and column of index 2
arrayInd(2, dim(matrix(seq(1:9), nrow = 3, ncol = 3, byrow = T)))

}
\seealso{
Other mathematics: 
\code{\link{logistic}()},
\code{\link{logit}()},
\code{\link{nearest}()}
}
\concept{mathematics}
