% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characteristic_contours.R
\name{candidate_contours}
\alias{candidate_contours}
\title{Identify candidates for the characteristic contour}
\usage{
candidate_contours(exam_curvature)
}
\arguments{
\item{exam_curvature}{A data frame with the same structure as \link{sample_curvature},
containing one row for each curvature \code{measurement}.}
}
\value{
An \emph{n}-row data frame containing one row for each candidate contour. Each
row will contain the \link{join_fields} and the following fields:

\describe{
\item{surface}{The corneal surface. Either \code{front} or \code{back}.}
\item{contour}{The dioptric power of the candidate contour.}
\item{segments}{The number of segments (polygons) that comprise the contour.}
\item{r_max}{The radial distance of K-max from the apex.}
\item{r_next}{The radial distance of K-next from the apex.}
}
}
\description{
Identifies the contours of an exam that are candidates for the
characteristic contour.
}
\details{
The underlying algorithm operates at the segment level. If any contour has a
segment that contains both K-max and K-next, only contours with segments
containing both points are considered candidates. If no contour has a segment
that contains both points (or K-next doesn't exist), contours with segments
containing a single point are considered candidates. In the unlikely case
that no contours have segments containing points, all contours are
considered candidates.
}
\examples{
candidate_contours(sample_curvature)

}
\seealso{
Other Characteristic Contours: 
\code{\link{contour_context}()}
}
\concept{Characteristic Contours}
