% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical_shapes.R
\docType{data}
\name{canonical_shapes}
\alias{canonical_shapes}
\title{canonical_shapes}
\format{
A data frame with the following columns:
\describe{
\item{last_name}{Last name of the patient}
\item{first_name}{First name of the patient}
\item{pat_id}{Patient ID}
\item{exam_date}{Date of the exam}
\item{exam_time}{Time of the exam}
\item{exam_eye}{'RIGHT' or 'LEFT'}
\item{cluster}{Oridinal identifier of the cluster}
\item{shape}{Name of the cluster}
}
}
\usage{
canonical_shapes
}
\description{
A data frame identifying the canonical shapes.
}
\keyword{datasets}
