% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characteristic_contours.R
\name{closest_silhouette}
\alias{closest_silhouette}
\title{Identify the closest canonical silhouette}
\usage{
closest_silhouette(exam_curvature, exam_power, exam_astig)
}
\arguments{
\item{exam_curvature}{A data frame containing one row for each curvature \code{measurement} and the
same columns as \link{sample_curvature}.}

\item{exam_power}{A number identifying the power of the exam's contour. Valid values for an
exam can be determined by calling \link{get_contour_levels}.}

\item{exam_astig}{A number, the angle (in degrees) of the exam's primary axis of astigmatism.}
}
\value{
A data frame containing the \link{join_fields}, corneal \code{surface} and
three additional columns: \code{contour}, the dioptric power of the exam's
contour, \code{closest_fit}, the index of the canonical silhouette that has
the highest percentage of overlap with the exam's silhouette, and
\code{average_distance}, the average percentage of overlap between the exam's
silhouette and the closest fitting member of the group.
}
\description{
Compares the silhouette of an exam's contour to the silhouettes of the
canonical shapes and identifies the canonical shape with the highest
percentage of overlap.
}
\details{
This function creates all of the necessary objects to perform the comparison.
Calling \link{IQeyes:silhouette_compare} or \link{IQeyes:silhouette_group_compare}
directly requires the creation of two objects for every contour, a silhouette
(filled shape) and a polygon. This function creates those objects for the
exam, but it requires the power of the contour and the angle of astigmatism
to do so. The silhouettes \link{IQeyes:canonical_silhouettes} and polygons
\link{IQeyes:canonical_polygons} for the canonical shapes are included as objects
in this package.
}
\examples{
closest_silhouette(sample_curvature, exam_power = 45.5, exam_astig = 33.6)

}
\seealso{
Other Characteristic Contours: 
\code{\link{candidate_contours}()},
\code{\link{contour_context}()}
}
\concept{Characteristic Contours}
