% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characteristic_contours.R
\name{contour_context}
\alias{contour_context}
\title{Identify candidates for characteristic contour}
\usage{
contour_context(exam_curvature, interp = F)
}
\arguments{
\item{exam_curvature}{A data frame with the same structure as \link{sample_curvature},
containing one row for each curvature \code{measurement}.}
}
\value{
A data frame, containing the output of \link{candidate_contours} and three
additional columns:

\describe{
\item{k_max_distance}{The number of other contours between a candidate and
K-max.}
\item{k_next_distance}{The number of other contours between a candidate and
K-next.}
\item{k_max_next_distance}{The difference between \code{k_max_distance} and
\code{k_next_distance}.}
}
}
\description{
Identifies the contours of an exam that are candidates for the
characteristic contour, and calculates the number of \emph{other}
contours that exist between a candidate contour and both K-max and K-next.
}
\details{
When trying to identify an exam's characteristic contour, the actual values
of K-max and K-next don't matter, but the relative context that they create
is important. This function quantifies that context by counting the number of
\emph{other} contours that exist between a candidate contour and both K-max
and K-next.

Internally, this function calls \link{k_next} and
\link{candidate_contours}.
}
\examples{
contour_context(sample_curvature, interp = T) |>
  dplyr::select(-tidyselect::all_of(join_fields))

}
\seealso{
Other characteristic contours: 
\code{\link{candidate_contours}()}
}
\concept{characteristic contours}
