% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contours.R
\name{contour_polygon}
\alias{contour_polygon}
\title{Convert a contour to a closed polygon}
\usage{
contour_polygon(exam_contour)
}
\arguments{
\item{exam_contour}{A data frame containing one row for each point of the contour and the same
columns as \link{sample_contour}. If the contour needs to be scaled and/or
rotated, the necessary transforms must be performed prior to calling
\code{contour_polygon}.}
}
\value{
A data frame containing the same columns as \code{exam_contour} with one
extra row for each segment of the contour.
}
\description{
Converts every segment of a contour to a closed polygon. The difference
between a \emph{segment} and a \emph{polygon} is that the latter contains two
points at the same location, representing the starting and ending points of
the polygon.
}
\details{
A contour is comprised of one or more segments. \code{contour_polygon}
takes the first point of each segment (based on its ordinal
position in the original data frame) and appends a replica to the end of the
the original data frame.
}
\examples{
contour_polygon(sample_contour) |> head()

}
\seealso{
\code{\link[=fill_contour]{fill_contour()}}, \code{\link[=scale_rotate]{scale_rotate()}}
\code{\link[=silhouette_compare]{silhouette_compare()}}, \code{\link[=silhouette_compare_group]{silhouette_compare_group()}}
}
