% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{contour_to_sf_polygon}
\alias{contour_to_sf_polygon}
\title{Convert a contour to a polygon}
\usage{
contour_to_sf_polygon(contour, required_points = 4)
}
\arguments{
\item{contour}{A data frame containing one row for each point of the contour and the same
columns as \link{sample_contour}. If the contour needs to be scaled and/or
rotated, the necessary transforms must be performed prior to calling
\code{contour_polygon}.}
}
\value{
An sf object.
}
\description{
Converts a contour to an sf polygon object.
}
\details{
If the contour contains segments that are fully enclosed in another segment,
this function will \emph{subtract} those segments from the resulting polygon.
}
\examples{
contour_to_sf_polygon(
  contour = get_contour(sample_curvature, contour_power = 44)
)

}
\seealso{
Other Polygons: 
\code{\link{silhouette_overlap}()},
\code{\link{silhouette_overlap_group}()},
\code{\link{silhouette_plot}()}
}
\concept{Polygons}
