% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{convert_name}
\alias{convert_name}
\title{Convert a Pentacam name to an R-friendly format}
\usage{
convert_name(name)
}
\arguments{
\item{name}{A character string containing the file or column name to be converted.}
}
\value{
A character string containing the converted name.
}
\description{
Converts all characters to lower case, replaces decimals and whitespace
characters with an underscore ("_"), replaces symbols with text (e.g., "°"
with "deg"), and more.
}
\examples{
convert_name('r4.0 324°')

}
\seealso{
Other utilities: 
\code{\link{parse_csv_to_list}()},
\code{\link{read_pentacam_csv}()}
}
\concept{utilities}
