% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_distributions.R
\name{d_mode}
\alias{d_mode}
\title{Dominant mode of a sample}
\usage{
d_mode(x)
}
\arguments{
\item{x}{A vector of numeric values.}
}
\value{
Returns the mode of \code{x}.
}
\description{
Calculates the mode for a vector of numeric values \code{x}.
}
\examples{
d_mode(rnorm(1000, mean = 2))

}
