% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_systems.R
\name{deg_to_rad}
\alias{deg_to_rad}
\title{Convert an angle from degrees to radians}
\usage{
deg_to_rad(angle_deg)
}
\arguments{
\item{angle_deg}{The angle in degrees.}
}
\value{
The angle in radians.
}
\description{
Convert an angle from degrees to radians.
}
\examples{
deg_to_rad(180)

}
\seealso{
\code{\link[=rad_to_deg]{rad_to_deg()}}
}
