% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_systems.R
\name{euclidean_dist}
\alias{euclidean_dist}
\title{Euclidean distance}
\usage{
euclidean_dist(x, y)
}
\arguments{
\item{x}{The \emph{x}-axis component.}

\item{y}{The \emph{y}-axis component.}
}
\value{
The distance of the point from the origin.
}
\description{
Calculates the Euclidean distance of a point (\emph{x}, \emph{y}) from the
origin (0, 0).
}
\examples{
euclidean_dist(2, 2)

}
