% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contours.R
\name{get_contour}
\alias{get_contour}
\title{Get contour}
\usage{
get_contour(exam_curvature, add_edge = F, contour_power)
}
\arguments{
\item{exam_curvature}{A data frame containing one row for each curvature \code{measurement} and the
same columns as \link{sample_curvature}.}

\item{add_edge}{A Boolean. If the contour extends beyond the edge of the scanned cornea,
\code{TRUE} will add an edge to close the contour. The \code{contour_id} for
the edge's segment will be \code{0}. If the contour doesn't extend beyond the
edge of the cornea, no edge will be added.}

\item{contour_power}{A number identifying the power of the contour to return.}
}
\value{
A data frame containing one row for each point of the contour and the same
columns as \link{sample_contour}.
}
\description{
Returns the full range of contours and the number of segments needed to plot
\code{exam_curvature} on a curvature map.
}
\details{
A contour is comprised of one or more segments, uniquely identified by
\code{contour_id}.

This function doesn't currently work for posterior surfaces because the
\code{plot_scale} function presumes the color scale is on the "absolute
scale", which is designed for the dioptric power of anterior surfaces.
}
\examples{
get_contour(sample_curvature, contour_power = 42) |>
  head()

}
\seealso{
Other Contours: 
\code{\link{contour_polygon}()},
\code{\link{fill_contour}()},
\code{\link{get_contour_levels}()},
\code{\link{scale_rotate}()},
\code{\link{silhouette_compare}()},
\code{\link{silhouette_compare_group}()}
}
\concept{Contours}
