% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_systems.R
\name{hausdorff_distance}
\alias{hausdorff_distance}
\title{Hausdorff distance}
\usage{
hausdorff_distance(contour_A, contour_B)
}
\arguments{
\item{contour_A}{A data frame containing contour A.}

\item{contour_B}{A data frame containing contour B.}
}
\value{
The Hausdorff distance between the two contours.
}
\description{
Calculates the Hausdorff distance between two contours.
}
\examples{
hausdorff_distance(
  contour_A = get_contour(sample_curvature, contour_power = 44),
  contour_B = get_contour(sample_curvature, contour_power = 44.5)
)

}
\seealso{
Other Coordinate Systems: 
\code{\link{cartesian_degrees}()},
\code{\link{cartesian_to_polar}()},
\code{\link{deg_to_rad}()},
\code{\link{euclidean_dist}()},
\code{\link{interpolate_measurements}()},
\code{\link{polar_to_cartesian}()},
\code{\link{quadrant_eye}()},
\code{\link{rad_to_deg}()},
\code{\link{rotate_shape}()},
\code{\link{signed_smallest_angle}()},
\code{\link{smallest_angle}()},
\code{\link{within_360}()}
}
\concept{Coordinate Systems}
