% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_systems.R
\name{interpolate_measurements}
\alias{interpolate_measurements}
\title{Interpolate measurements over an expanded grid}
\usage{
interpolate_measurements(source_dat, ...)
}
\arguments{
\item{source_dat}{A data frame containing one row for each \code{measurement}. Each row must contain
at least three columns: a set of Cartesian coordinates (\code{x} and
\code{y}) and a \code{measurement}.}

\item{...}{Additional parameters that get passed to \code{akima::interp()}}
}
\value{
A new data frame with the interpolated measurements.
}
\description{
Interpolates measurements over an expanded
grid of Cartesian coordinates (\emph{x} and \emph{y}).
}
\details{
The \emph{x} and \emph{y} axes of the expanded grid will contain the same
number of unique values as original axes, but they will be equally spaced
along those axes.

The interpolated measurements will \emph{not} necessarily include the
original measurements.
}
\examples{
interpolate_measurements(sample_curvature[1:10, c('x', 'y', 'measurement')])

}
