% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iq_reference_stats.R
\docType{data}
\name{iq_reference_stats}
\alias{iq_reference_stats}
\title{iq_reference_stats}
\format{
A data frame with the following columns:
\describe{
\item{index}{Name of the index}
\item{mean}{Mean of the index}
\item{sd}{Standard deviation of the index}
\item{min}{Minimum of the index}
\item{max}{Minimum of the index}
\item{population}{Either "normal" or "abnormal"}
\item{doi}{Digital Object Identifier}
}
}
\usage{
iq_reference_stats
}
\description{
Reference statistics used by IQeyes.
}
\keyword{datasets}
