% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_next.R
\name{k_next}
\alias{k_next}
\title{Find the peaks on a curvature map}
\usage{
k_next(curvature_m, ignore_singletons = T)
}
\arguments{
\item{curvature_m}{A square numeric matrix. The rows must be labeled with their respective
\emph{y}-axis positions, and the columns with their respective \emph{x}-axis
positions. The values of the matrix must be the dioptric power at each
(\emph{x}, \emph{y}) position.}

\item{ignore_singletons}{A Boolean. \code{TRUE} to ignore peaks that contain just one point.}
}
\value{
Returns a data frame that contains one row for each peak on the curvature
map. The first few columns of the data frame will contain the
\code{join_fields}. The remaining columns describe the characteristics of
each peak:

\describe{
\item{peak_id}{The ordinal of the peak, ranked by the dioptric power of the
peak's apex. The row with \code{peak_id == 1} will be K-max and the row with
\code{peak_id == 2} will be K-next.}
\item{x}{The peak's \emph{x}-axis position. See details.}
\item{y}{The peak's \emph{y}-axis position.}
\item{power}{The dioptric power of the peak.}
\item{pct_coverage}{The percentage of points on the cornea that descend
from the peak.}
\item{r}{The radial distance of the peak from the vertex (in mm).}
\item{theta}{The angle of the peak from the 3 o'clock position (in
radians).}
\item{angle}{The angle of the peak from the 3 o'clock position (in
degrees).}
\item{alignment_deg}{The angle (in degrees) between a peak and the
projection of K-max in its opposite direction. Quantifies the departure from
symmetry. When K-max and another peak (particularly K-next) are
symmetrically aligned, their radial axes are 180° apart. As such,
\code{alignment_deg} is calculated as 180° \emph{minus} the smallest angle,
so when K-max and another peak are 180° apart, the result is 0°. Generally,
the smallest angle between the K-max and another peak will be less than
180°, and the departure from symmetry will be larger.}
\item{distance_to_max}{The Euclidean distance (in mm) between the peak and
another peak.}
}
}
\description{
Finds the points of maximum dioptric power on a curvature map, including
K-max, K-next, and any other peaks.
}
\details{
Oftentimes, a peak is comprised of several points with the same dioptric
power. When that happens, the (\emph{x}, \emph{y}) position of the peak is
calculated as the centroid of those points.
}
\examples{
interpolate_measurements(sample_curvature) |>
  dplyr::mutate(power = anterior_power(z)) |>
  dplyr::select(x, y, power) |>
  reshape2::acast(y ~ x, value.var = 'power') |>
  k_next()

}
\seealso{
Other K-next: 
\code{\link{antipodal_power}()},
\code{\link{find_peaks}()},
\code{\link{k_next_old}()}
}
\concept{K-next}
