% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvature.R
\name{kappa_from_radius}
\alias{kappa_from_radius}
\title{Kappa from the radius of curvature}
\usage{
kappa_from_radius(R)
}
\arguments{
\item{R}{Radius of curvature.}
}
\value{
Returns the curvature.
}
\description{
Calculate curvature (kappa, K) from the radius of curvature (R).
}
\details{
Curvature (K) is different from the conic constant (Q)
}
\examples{
kappa_from_radius(7)

}
\references{
\emph{Pentacam Interpretation Guide}, p 7.
}
