% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_distributions.R
\name{knee}
\alias{knee}
\title{Find the knee in a scree plot}
\usage{
knee(df)
}
\arguments{
\item{df}{A data frame that contains the \emph{x} and \emph{y} coordinates of the points
that comprise a scree plot, ordered from small to large values of \emph{x}.
The columns must be labeled \code{x} and \code{y}.}
}
\value{
Returns the row of the data frame that contains the knee of the curve.
}
\description{
Finds the knee in the curve of a two-dimensional scree plot.
}
\examples{
data.frame(x = 1:15, y = sort(rpois(15, 42), decreasing = T)) |>
  knee()

}
