% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematics.R
\name{logistic}
\alias{logistic}
\title{Compute the logistic transformation of number}
\usage{
logistic(x)
}
\arguments{
\item{x}{A numeric value.}
}
\value{
Returns \code{1 / (1 + exp(-x))}.
}
\description{
Maps a numeric value from [−Inf, +Inf] to [0, 1]. Useful for mapping real
numbers to a probability.
}
\details{
Inverse of the logit function.
}
\examples{
logistic(x = 64)
logistic(Inf)

}
\seealso{
\code{\link[=logit]{logit()}}
}
