% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematics.R
\name{logit}
\alias{logit}
\title{Compute the logit transformation of proportions or percentages}
\usage{
logit(p)
}
\arguments{
\item{p}{A numeric probability ranging from 0 to 1.}
}
\value{
Returns \code{log(p / (1 - p))}.
}
\description{
Maps a probability value from [0, 1] to a real number in [−Inf, +Inf].
Useful for mapping a probability to a real number.
}
\details{
Inverse of the logistic function.
}
\examples{
logit(p = 0.5)
logit(1/2)

}
\seealso{
\code{\link[=logistic]{logistic()}}

Other mathematics: 
\code{\link{array_position}()},
\code{\link{logistic}()},
\code{\link{nearest}()}
}
\concept{mathematics}
