% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_distributions.R
\name{mean_half_normal}
\alias{mean_half_normal}
\title{Mean of a half-normal distribution}
\usage{
mean_half_normal(x, ...)
}
\arguments{
\item{x}{A numeric vector of values from a half-normal distribution.}
}
\value{
Returns the mean of \code{x}.
}
\description{
Calculates the mean (expected value) of a sample from a half-normal
distribution.
}
\details{
\code{mean_half_normal} adjusts \code{mean(x)} to account for the characteristics
of the half-normal distribution
}
\examples{
mean_half_normal(abs(rnorm(10)))

}
\seealso{
\code{\link[=sd_half_normal]{sd_half_normal()}}, \code{\link[=mean_folded_normal]{mean_folded_normal()}}, \code{\link[=sd_folded_normal]{sd_folded_normal()}}
}
