% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{parse_csv_to_list}
\alias{parse_csv_to_list}
\title{Read a detailed Pentacam CSV into a list}
\usage{
parse_csv_to_list(file_name, delimiter = ",", pachy_flag = F)
}
\arguments{
\item{file_name}{A character string containing the name of the file name to read.}

\item{delimiter}{A character identifying the delimiter that the Pentacam file uses to separate
fields/columns.}

\item{pachy_flag}{A Boolean. \code{TRUE} to indicate that the file contains pachymetry data.}
}
\value{
A list containing the file that has been read.
}
\description{
Reads one the Pentacam's exported CSV files into a list object. The list will
contain one element for each section of the file.
}
\details{
The date and time fields in the \code{patient} and \code{examination}
sections are \emph{not} converted to lubridate date and time types.

The section \code{cornea edge} contains a bug.
}
\seealso{
Other Utilities: 
\code{\link{convert_name}()},
\code{\link{read_pentacam_csv}()}
}
\concept{Utilities}
