% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_distributions.R
\name{pct_to_z}
\alias{pct_to_z}
\title{Z-score from a percentage (of a normal distribution)}
\usage{
pct_to_z(pct, two_tail = T)
}
\arguments{
\item{pct}{Percentage of values.}

\item{two_tail}{A Boolean indicating whether the z-score should be based on two (\code{T}) or
one (\code{F}) tails. Basically, two_tail = \code{T} cuts off values outside
\code{±z * σ}.}
}
\value{
Returns the z-score for the percentage of values.
}
\description{
Calculates a z-score based on the percentage of values from a normal
distribution.
}
\examples{
pct_to_z(pct = 0.5, two_tail = F)
pct_to_z(c(-0.5, 0.5), two_tail = T)

}
\seealso{
Other statistical functions: 
\code{\link{d_mode}()},
\code{\link{knee}()},
\code{\link{mean_folded_normal}()},
\code{\link{mean_half_normal}()},
\code{\link{sd_folded_normal}()},
\code{\link{sd_half_normal}()},
\code{\link{z_score}()},
\code{\link{z_to_pct}()}
}
\concept{statistical functions}
