% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_matrix}
\alias{plot_matrix}
\title{Plot the detailed matrix data}
\usage{
plot_matrix(
  exam_matrix,
  color_scale = c("elevation_colors", "pachymetry_colors", "curvature_colors"),
  scale_units = "Measurement",
  invert_order = F,
  truncate_circle = T,
  circle_dia = 9,
  axes = T,
  legend = T,
  show_arrows = F,
  arrows_legend = T,
  theme_minimal = T,
  coord_fixed = T
)
}
\arguments{
\item{exam_matrix}{A numeric matrix containing the values to plot. The row and column names of
the matrix must contain y- and x-axis positions, respectively, of the z-axis
values.}

\item{color_scale}{A data frame containing the \code{breaks} and \code{hex_color} values for the
contours of the plot.}

\item{scale_units}{A character string containing the units of measure for the z-axis values.
Used to label the legend.}

\item{invert_order}{A Boolean. \code{FALSE} (the default) to retain the order of the
\code{color_scale}. \code{TRUE} to invert the order.}

\item{truncate_circle}{A Boolean. \code{TRUE} to limit the plotted region of the matrix to a circle
of diameter \code{circle_dia}.}

\item{circle_dia}{A numeric value. When \code{truncate_circle} is \code{TRUE}, the diameter of
the plotted region of the matrix.}

\item{axes}{A Boolean. \code{TRUE} to display the \emph{x} and \emph{y} axes on the plot.}

\item{legend}{A Boolean. \code{TRUE} to display a legend/key.}

\item{show_arrows}{A Boolean. \code{TRUE} to display the point-to-point change in z-axis values
as arrows.}

\item{arrows_legend}{A Boolean. \code{TRUE} to display a legend/key for arrows.}

\item{theme_minimal}{A Boolean. \code{TRUE} to use ggplot's minimal theme (e.g., no background
color).}

\item{coord_fixed}{A Boolean. \code{TRUE} to ensure the \emph{x} and \emph{y} axes remain fixed
(i.e., equal), regardless of differences in the extents of the \emph{x} and
\emph{y} axes.}
}
\value{
A ggplot object.
}
\description{
Returns (but does not render) a filled contour plot of a matrix dataset.
}
\details{
To plot anterior axial/sagittal curvature, the z-axis values must be in
diopters. Use \link{anterior_power} to convert radii values to diopters.

Posterior curvature has not yet been implemented: It requires only the
definition of a color scale.

Tangential curvature has not yet been implemented.
}
\seealso{
Other plots: 
\code{\link{curvature_plot}()},
\code{\link{plot_scale}()}
}
\concept{plots}
