% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_scale}
\alias{plot_scale}
\title{Contour breaks needed to plot an anterior curvature map}
\usage{
plot_scale(z)
}
\arguments{
\item{z}{A numeric vector containing power (in diopters). Currently
works for \emph{anterior} surfaces only.}
}
\value{
A numeric vector containing the breaks on the \emph{z}-axis scale (in
diopters).
}
\description{
Returns the subset of contour breaks from \code{absolute_scale} required to
plot a curvature map of the values in \code{z}.
}
\details{
The contour levels on a curvature map are defined by the breaks in
\code{absolute_scale}, which is based on the Pentacam's American Style scale
for the dioptric power of anterior curvature maps.
}
\examples{
sample_curvature$measurement |>
  anterior_power() |>
  plot_scale()

}
\seealso{
Other plots: 
\code{\link{curvature_plot}()},
\code{\link{plot_matrix}()}
}
\concept{plots}
