% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvature.R
\name{posterior_power}
\alias{posterior_power}
\title{Posterior power}
\usage{
posterior_power(R)
}
\arguments{
\item{R}{Rear refractive index.}
}
\value{
Returns the optical power of the posterior surface of a cornea.
}
\description{
Calculates the dioptric power of the posterior (rear) surface of a cornea
based on the radius of curvature.
}
\examples{
posterior_power(5)

}
\references{
\emph{Pentacam Interpretation Guide}, p 7.
}
\seealso{
Other curvature: 
\code{\link{anterior_power}()},
\code{\link{kappa_from_radius}()},
\code{\link{true_net_power}()}
}
\concept{curvature}
