% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_systems.R
\name{rad_to_deg}
\alias{rad_to_deg}
\title{Convert an angle from radians to degrees}
\usage{
rad_to_deg(angle_rad)
}
\arguments{
\item{angle_rad}{The angle in radians.}
}
\value{
The angle in degrees.
}
\description{
Convert an angle from radians to degrees.
}
\examples{
rad_to_deg(pi)

}
\seealso{
\code{\link[=deg_to_rad]{deg_to_rad()}}
}
