% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radii_degrees.R
\docType{data}
\name{radii_degrees}
\alias{radii_degrees}
\title{radii_degrees}
\format{
A data frame with five (5) columns:
\describe{
\item{column_name}{Column name}
\item{ring_diam}{Ring diameter where the measurement was taken}
\item{angle}{Angle where the measurement was taken}
\item{x}{x-axis coordinate of the measurement}
\item{y}{y-axis coordinate of the measurement}
}
}
\usage{
radii_degrees
}
\description{
A reference data frame, mapping the column names of curvature measurements
to polar and Cartesian coordinates.
}
\keyword{datasets}
