% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{read_pentacam_csv}
\alias{read_pentacam_csv}
\title{Read a Pentacam CSV file}
\usage{
read_pentacam_csv(
  file_name,
  delimiter = ";",
  keep_ok_only = T,
  keep_dup_rows = F,
  keep_empty_cols = F
)
}
\arguments{
\item{file_name}{A character string containing the name of the file name to read.}

\item{delimiter}{A character identifying the delimiter that the Pentacam file uses to separate
fields/columns.}

\item{keep_ok_only}{A Boolean. \code{TRUE} to return only records/rows with a Penatacam status
of "OK".}

\item{keep_dup_rows}{A Boolean. \code{TRUE} to return all records/rows, including duplicates.}

\item{keep_empty_cols}{A Boolean. \code{TRUE} to return all columns, even when none of the
records/rows contain values.}
}
\value{
A data frame containing the file that has been read.
}
\description{
Reads a Pentacam CSV file into a data frame. The function also converts the
exam date and time columns to lubridate date and period fields, respectively,
converts the patient's date of birth to a date field (and calculates their
age at the time of the exam), converts the exam eye to a factor, and converts
the file's column names to an R-friendly format.
}
\seealso{
Other Utilities: 
\code{\link{convert_name}()}
}
\concept{Utilities}
