% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_polygons.R
\docType{data}
\name{reference_polygons}
\alias{reference_polygons}
\title{reference_polygons}
\format{
A polygon object:
\describe{
\item{cluster}{Oridinal identifier of the cluster}
\item{geometry}{Geometry of the reference shape}
}
}
\usage{
reference_polygons
}
\description{
A polygon object containing all of the reference shapes
}
\keyword{datasets}
