% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_shapes.R
\docType{data}
\name{reference_shapes}
\alias{reference_shapes}
\title{reference_shapes}
\format{
A data frame with the following columns:
\describe{
\item{last_name}{Last name of the patient}
\item{first_name}{First name of the patient}
\item{pat_id}{Patient ID}
\item{exam_date}{Date of the exam}
\item{exam_time}{Time of the exam}
\item{exam_eye}{'RIGHT' or 'LEFT'}
\item{cluster}{Oridinal identifier of the cluster}
\item{shape}{Name of the cluster}
}
}
\usage{
reference_shapes
}
\description{
A data frame identifying the reference shapes.
}
\keyword{datasets}
