% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_astig.R
\docType{data}
\name{sample_astig}
\alias{sample_astig}
\title{sample_astig}
\format{
A data frame with the following columns:
\describe{
\item{last_name}{Last name of the patient}
\item{first_name}{First name of the patient}
\item{pat_id}{Patient ID}
\item{exam_date}{Date of the exam}
\item{exam_time}{Time of the exam}
\item{exam_eye}{'RIGHT' or 'LEFT'}
\item{ring_diam}{Ring diameter of measurement}
\item{ast}{Degree of astigmatism (in diopters)}
\item{axs}{Axis of astigmatism (in degress)}
}
}
\usage{
sample_astig
}
\description{
A sample exam's astigmatism data (from CHAMBER).
}
\keyword{datasets}
