% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_contour.R
\docType{data}
\name{sample_contour}
\alias{sample_contour}
\title{sample_contour}
\format{
A data frame with the following columns:
\describe{
\item{last_name}{Last name of the patient}
\item{first_name}{First name of the patient}
\item{pat_id}{Patient ID}
\item{exam_date}{Date of the exam}
\item{exam_time}{Time of the exam}
\item{surface}{'RIGHT' or 'LEFT'}
\item{exam_eye}{'FRONT' or 'BACK'}
\item{x}{x-axis coordinate of a point on the contour}
\item{y}{y-axis coordinate of a point on the contour}
\item{segment_id}{Ordinal identifier for each segment of a \emph{n}-segment contour}
\item{contour}{Dioptric power of the contour}
}
}
\usage{
sample_contour
}
\description{
A sample exam's contour data.
}
\keyword{datasets}
