% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_curvature.R
\docType{data}
\name{sample_curvature}
\alias{sample_curvature}
\title{sample_curvature}
\format{
A data frame with thirteen (13) elements:
\describe{
\item{last_name}{Last name of the patient}
\item{first_name}{First name of the patient}
\item{pat_id}{Patient ID}
\item{exam_date}{Date of the exam}
\item{exam_time}{Time of the exam}
\item{exam_eye}{'RIGHT' or 'LEFT'}
\item{surface}{'FRONT' or 'BACK' surface of the cornea}
\item{column_name}{Name of the column in COR-PWR that contained the measurement}
\item{measurement}{Measured curvature radius}
\item{ring_diam}{Ring diameter where the measurement was taken}
\item{angle}{Angle where the measurement was taken}
\item{x}{x-axis coordinate of the measurement}
\item{y}{y-axis coordinate of the measurement}
}
}
\usage{
sample_curvature
}
\description{
A sample exam's curvature data (from COR-PWR).
}
\keyword{datasets}
