% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contours.R
\name{scale_rotate}
\alias{scale_rotate}
\title{Scale and rotate}
\usage{
scale_rotate(shape, axs = 0, r_target = 4)
}
\arguments{
\item{shape}{A data frame containing one row for each point that defines a shape (e.g.,
a silhouette or contour). Currently, \code{scale_rotate()} expects the
data frame to contain the join fields.}

\item{axs}{A number. Oftentimes, this should be the angle (in degrees) of the primary
axis of astigmatism for an exam.}

\item{r_target}{A number defining the target radial length of the scaled shape.}
}
\value{
A data frame containing the same columns as \code{shape} with transformed
\emph{x} and \emph{y} coordinates. No other values are changed.
}
\description{
Scales and rotates a shape defined by a set of \emph{x} and \emph{y}
coordinates. Can be either a curvature (e.g., \link{sample_curvature}) or
contour data frame (e.g., \link{sample_contour}).
}
\details{
The angle of rotation is determined by the primary axis of astigmatism for
the exam. If \code{axs} > 90, its value is shifted by -180°, ensuring that
the contour isn't flipped along the horizontal axis.

The scaling factor is determined by the max radial length of any point on the
shape, not the end-to-end length of the shape, which is especially
important for asymmetric shapes.
}
\examples{
scale_rotate(sample_contour, axs = 36.2) |>
  head()
scale_rotate(sample_curvature, axs = 36.2) |>
  head()

}
\seealso{
Other Contours: 
\code{\link{contour_polygon}()},
\code{\link{fill_contour}()},
\code{\link{get_contour}()},
\code{\link{get_contour_levels}()},
\code{\link{silhouette_compare}()},
\code{\link{silhouette_compare_group}()}
}
\concept{Contours}
