% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_distributions.R
\name{sd_folded_normal}
\alias{sd_folded_normal}
\title{Standard deviation of a folded-normal distribution}
\usage{
sd_folded_normal(mu, sigma)
}
\arguments{
\item{mu}{Mean of the normal distribution.}

\item{sigma}{Standard deviation of the normal distribution.}
}
\value{
Returns the standard deviation.
}
\description{
Calculates the standard deviation of a folded-normal distribution based on
the parameters of a normal distribution. Presumes that the folded-normal
distribution is the absolute value of the normal distribution.
}
\examples{
sd_folded_normal(mu = 3, sigma = 1)

}
\seealso{
\code{\link[=mean_folded_normal]{mean_folded_normal()}}, \code{\link[=sd_half_normal]{sd_half_normal()}}, \code{\link[=mean_half_normal]{mean_half_normal()}}
}
