% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_distributions.R
\name{sd_half_normal}
\alias{sd_half_normal}
\title{Standard deviation of a half-normal distribution}
\usage{
sd_half_normal(x, ...)
}
\arguments{
\item{x}{A numeric vector of values from a half-normal distribution.}
}
\value{
Returns the standard deviation of \code{x}.
}
\description{
Calculates the standard deviation of sample from a half-normal distribution.
}
\details{
\code{sd_half_normal} adjusts \code{sd(x)} to account for the characteristics
of the half-normal distribution
}
\examples{
sd_half_normal(abs(rnorm(10)))

}
\seealso{
\code{\link[=mean_half_normal]{mean_half_normal()}}, \code{\link[=sd_folded_normal]{sd_folded_normal()}}, \code{\link[=mean_folded_normal]{mean_folded_normal()}}
}
