% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematics.R
\name{signed_pct_diff}
\alias{signed_pct_diff}
\title{Signed percentage difference}
\usage{
signed_pct_diff(x, y)
}
\arguments{
\item{x}{A number. If \code{x} is 0, \code{y} needs to be non-zero.}

\item{y}{A number. If \code{y} is 0, \code{x} needs to be non-zero.}
}
\value{
Returns \code{(x - y) / (max(abs(x), abs(y)))}. Note
that the result is directional: If \code{x} > \code{y}, then
\code{signed_pct_diff} returns a positive value; If \code{x} < \code{y}, then
\code{signed_pct_diff} returns a negative value.
}
\description{
Calculates the signed percentage difference between two numbers.
}
\examples{
signed_pct_diff(x = 5, y = 10)
signed_pct_diff(5, 10)

}
