% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contours.R
\name{silhouette_compare}
\alias{silhouette_compare}
\title{Compare the silhouettes of two contours}
\usage{
silhouette_compare(shape_A, polygon_A, shape_B, polygon_B, show_plots = F)
}
\arguments{
\item{shape_A}{A data frame containing the filled shape for contour A.}

\item{polygon_A}{A data frame containing the polygon object for contour A.}

\item{shape_B}{A data frame containing the filled shape for contour B.}

\item{polygon_B}{A data frame containing the polygon object for contour B.}

\item{show_plots}{A Boolean. \code{TRUE} to render two plots that illustrate comparisons being
made.}
}
\value{
The percentage of overlap between the two silhouettes, averaged over both
directions.
}
\description{
Each of silhouettes is defined by two objects, a filled shape and a polygon.
These objects must be created before calling \code{silhouette_compare()}. The
filled shapes are created by calling \code{fill_contour()}, and polygon
objects are created by calling \code{contour_polygon()}.
}
\examples{
silhouette_compare(
  shape_A = fill_contour(sample_curvature, sample_contour),
  polygon_A = contour_polygon(sample_contour),
  shape_B = fill_contour(sample_curvature, sample_contour),
  polygon_B = contour_polygon(sample_contour)
)

}
\seealso{
\code{\link[=fill_contour]{fill_contour()}}, \code{\link[=contour_polygon]{contour_polygon()}},
\code{\link[=scale_rotate]{scale_rotate()}}, \code{\link[=silhouette_compare_group]{silhouette_compare_group()}}
}
