% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contours.R
\name{silhouette_plot_old}
\alias{silhouette_plot_old}
\title{Plot the overlap of one silhouette with a group of silhouettes}
\usage{
silhouette_plot_old(ref_polygon, group_polygons)
}
\arguments{
\item{ref_polygon}{A data frame containing the polygon object for the exam.}

\item{group_polygons}{A data frame containing polygons for the group of contours. The data frame
\emph{must} include a \code{cluster} column that uniquely identifies
each member of the group. The cluster IDs for individual contours (i.e.,
members of the group) must be the same in both \code{group_shapes} and
\code{group_polygons}.

@details
For convenience, the silhouettes \link{IQeyes:canonical_silhouettes} and polygons
\link{IQeyes:canonical_polygons} for the canonical shapes are included as
objects in this package.

For custom groups, the data frame passed into \code{group_polygons} requires
a \code{cluster} column, which can be passed into and returned by
\link{contour_polygon}.}
}
\value{
A ggplot2 object that that shows the overlapping silhouettes.
}
\description{
Plot the silhouette of one exam's contour with the silhouettes of a group.

Plotting requires a polygon (an outline with connected ends) for each of the
shapes. These objects must be created by calling \link{contour_polygon}
before calling \code{silhouette_plot_old()}.
}
\examples{
sample_polygon <- get_contour(sample_curvature, add_edge = T, contour_power = 45.5) |>
  scale_rotate(axs = 33.6) |>
  contour_polygon()

silhouette_plot_old(
  sample_polygon,
  contour_polygon(get_contour(sample_curvature, contour_power = 45))
)

silhouette_plot_old(
  ref_polygon = sample_polygon,
  group_polygons = canonical_polygons
)

}
\seealso{
Other Contours: 
\code{\link{contour_polygon}()},
\code{\link{fill_contour}()},
\code{\link{get_contour}()},
\code{\link{get_contour_levels}()},
\code{\link{scale_rotate}()},
\code{\link{silhouette_compare}()},
\code{\link{silhouette_compare_group}()}
}
\concept{Contours}
