% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_systems.R
\name{smallest_angle}
\alias{smallest_angle}
\title{The smallest angle (in degrees) between two radial lines}
\usage{
smallest_angle(angle_1, angle_2)
}
\arguments{
\item{angle_1}{The angle (in degrees) of the first radial line.}

\item{angle_2}{The angle (in degrees) of the second radial line.}
}
\value{
The angle between with the two lines.
}
\description{
Calculates the smallest angle (in degrees) between two radial lines.
}
\details{
The result is essentially the absolute difference between two angles.
}
\seealso{
\code{\link[=signed_smallest_angle]{signed_smallest_angle()}}
#' @examples
smallest_angle(45, 90)
smallest_angle(90, 45)
}
