% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meridians.R
\name{srax}
\alias{srax}
\title{Calculate the skewed radial axis (SRAX) for an exam}
\usage{
srax(exam_meridians, radius = 1.5)
}
\arguments{
\item{exam_meridians}{A data frame with the same structure as the output of
\link{curvature_meridians}.}

\item{radius}{A number identifying the radius of the axes to use for the skew calculation.}
}
\value{
A data frame containing the \code{join_fields} and \code{srax}, the skew of
radial axes.
}
\description{
The SRAX is 180 minus the smallest angle between the steepest radial axes
"above" and "below" the flat axis of astigmatism.
}
\details{
The formal definition of SRAX uses the horizontal meridian to split the
cornea into hemispheres. Presuming that \link{curvature_meridians} is used
to identify the steepest radial axes, this function uses the flat axis of
astigmatism identified by the Pentacam at the 3, 5, or 7 mm ring.
}
\examples{
curvature_meridians(sample_curvature, sample_astig) |>
  srax()

}
\references{
Rabinowitz, Yaron S. "Videokeratographic Indices to Aid in Screening for
Keratoconus." \emph{Journal of Refractive Surgery}, 1995 Sep-Oct;11(5):371-9.
\url{https://doi.org/10.3928/1081-597X-19950901-14}.
}
\seealso{
Other meridians: 
\code{\link{curvature_meridians}()}
}
\concept{meridians}
