% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_systems.R
\name{within_2pi}
\alias{within_2pi}
\title{Keep an angle within the 0--2π range}
\usage{
within_2pi(angle)
}
\arguments{
\item{angle}{An angle in radians.}
}
\value{
The equivalent angle within the 0--2π range.
}
\description{
Converts any angle (in radians) to the equivalent angle in the 0--2π range.
}
\details{
Adding or subtracting two angles can sometimes result in a new angle that is
outside the 0--2π range. This function converts the result to the equivalent
angle within the 0--2π range.
}
\examples{
within_2pi(2 * pi)
within_2pi(pi)
within_2pi(-pi)

}
\seealso{
Other Coordinate Systems: 
\code{\link{cartesian_degrees}()},
\code{\link{cartesian_to_polar}()},
\code{\link{deg_to_rad}()},
\code{\link{euclidean_dist}()},
\code{\link{hausdorff_distance}()},
\code{\link{interpolate_measurements}()},
\code{\link{polar_to_cartesian}()},
\code{\link{quadrant_eye}()},
\code{\link{rad_to_deg}()},
\code{\link{rotate_shape}()},
\code{\link{signed_smallest_angle}()},
\code{\link{smallest_angle}()},
\code{\link{within_360}()}
}
\concept{Coordinate Systems}
