% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_systems.R
\name{within_360}
\alias{within_360}
\title{Keep an angle within the 0--360° range}
\usage{
within_360(x)
}
\arguments{
\item{x}{An angle within the -360--720° range.}
}
\value{
The same angle within the 0--360° range.
}
\description{
Converts an angle from a -360--720° range to a 0--360° range.
}
\details{
Adding or subtracting two angles can sometimes result in a new angle that is
outside the 0--360° range. This function converts the result to the same
angle within the 0--360° range.
}
\examples{
within_360(551)
within_360(42)
within_360(-22)
}
