% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_distributions.R
\name{z_score}
\alias{z_score}
\title{Z-score}
\usage{
z_score(x, mean, sd)
}
\arguments{
\item{x}{A value from a normal distribution.}

\item{mean}{The mean of the normal distribution}

\item{sd}{The standard deviation of the normal distribution}
}
\value{
Returns the z-score of \code{x}.
}
\description{
Calculates the z-score for a value \code{x} taken from a normal distribution
defined by it mean and standard deviation.
}
\examples{
z_score(0.5, mean = 0, sd = 1)

}
\seealso{
\code{\link[=pct_to_z]{pct_to_z()}}, \code{\link[=z_to_pct]{z_to_pct()}}
}
