% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_distributions.R
\name{z_to_pct}
\alias{z_to_pct}
\title{Z-score to a percentage (of a normal distribution)}
\usage{
z_to_pct(z, two_tail = T)
}
\arguments{
\item{z}{Mean of a normal distribution.}

\item{two_tail}{A Boolean indicating whether the z-score is based on two (\code{T}) or
one (\code{F}) tails. Basically, two_tail = \code{T} cuts off values outside
\code{±z * σ}.}
}
\value{
Returns the percentage of a normal distribution.
}
\description{
Converts a z-score to the percentage of a normal distribution that
lie within z standard deviations of the mean.
}
\examples{
z_to_pct(z = 3, two_tail = T)
z_to_pct(c(1, 2, 3), two_tail = F)

}
\seealso{
\code{\link[=pct_to_z]{pct_to_z()}}, \code{\link[=z_score]{z_score()}}
}
