% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_next.R
\name{antipodal_power}
\alias{antipodal_power}
\title{Dioptric power at the antipode of K-max}
\usage{
antipodal_power(exam_curvature, exam_k_max, interp = F)
}
\arguments{
\item{exam_curvature}{A data frame containing, at a minimum, the \code{x}- and \code{y}-axis
position of each curvature (i.e., radius of curvature) \code{measurement}.
One row for each \code{measurement}.}

\item{exam_k_max}{A data frame containing the \link{join_fields} and the \code{x}, \code{y},
and \code{measurement} of K-max.}

\item{interp}{A Boolean. \code{TRUE} to interpolate measurements. See details.}
}
\value{
A one-row data frame containing the \link{join_fields} and the following
columns:

\describe{
\item{k_max}{Dioptric power of K-max.}
\item{opposite_power}{Dioptric power of the antipode.}
\item{x}{The \emph{x}-axis position of the measurement. See details.}
\item{y}{The \emph{y}-axis position of the measurement.}
\item{opposite_x}{The \emph{x}-axis position of the antipode. See details.}
\item{opposite_y}{The \emph{y}-axis position of the antipode.}
\item{euch_dist}{Euchlidean distance of the selected \code{measurement} from
the location of "exact" antipode of K-max. See details.}
}
}
\description{
Returns the dioptric power at the antipode of K-max.
}
\details{
The antipode of K-max is the point directly opposite of K-max on the
curvature map. Since measurement data for that \emph{exact} position is
rarely available, this function finds the point that's nearest to the
antipode and returns its position and power.

The \link{join_fields} in the returned data frame come from the
\code{exam_k_max}.
}
\seealso{
Other K-next: 
\code{\link{find_peaks}()},
\code{\link{k_next}()},
\code{\link{k_next_old}()}
}
\concept{K-next}
