% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characteristic_contours.R
\name{candidate_contours}
\alias{candidate_contours}
\title{Identify candidates for the characteristic contour}
\usage{
candidate_contours(exam_curvature)
}
\arguments{
\item{exam_curvature}{A data frame containing one row for each curvature \code{measurement} and the
same columns as \link{sample_curvature}.}
}
\value{
An emph{n}-row data frame containing one row for each candidate contour. Each
row will contain the \link{join_fields}, the cornea \code{surface}, and
the dioptric power of the candidate \code{contour}. In addition, each row
will contain the number of \code{segments} that comprise the contour, and the
radial distance of K-max(\code{r_max}) and K-next (\code{r_next}) from the
apex of the curvature map.
}
\description{
Identifies the contours of an exam that are candidates for the
characteristic contour.
}
\details{
The underlying algorithm operates at the segment level. If any contour has a
segment that contains both K-max and K-next, only contours with segments
containing both points are considered candidates. If no contour has a segment
that contains both points (or K-next doesn't exist), contours with segments
containing a single point are considered candidates. In the unlikely case
that no contours have any segments containing any points, all contours are
considered candidates.
}
\examples{
candidate_contours(sample_curvature)

}
\seealso{
Other Characteristic Contours: 
\code{\link{closest_silhouette}()},
\code{\link{contour_context}()}
}
\concept{Characteristic Contours}
