% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_systems.R
\name{cartesian_degrees}
\alias{cartesian_degrees}
\title{Convert the Cartesian coordinates of a point to an angle (in degrees)}
\usage{
cartesian_degrees(x, y)
}
\arguments{
\item{x}{The \emph{x}-axis component.}

\item{y}{The \emph{y}-axis component.}
}
\value{
the angle (\emph{theta}) in degrees.
}
\description{
Calculates the angle (in degrees) formed by the \emph{x}- and
\emph{y}-coordinates of a point on a Cartesian plane.
}
\examples{
cartesian_degrees(2, 2)
}
\seealso{
\code{\link[=cartesian_to_polar]{cartesian_to_polar()}}
}
