% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_systems.R
\name{cartesian_degrees}
\alias{cartesian_degrees}
\title{Convert the Cartesian coordinates of a point to an angle (in degrees)}
\usage{
cartesian_degrees(x, y)
}
\arguments{
\item{x}{The \emph{x}-axis component.}

\item{y}{The \emph{y}-axis component.}
}
\value{
The angle (\emph{theta}) in degrees.
}
\description{
Calculates the angle (in degrees) formed by the \emph{x}- and
\emph{y}-coordinates of a point on a Cartesian plane.
}
\examples{
cartesian_degrees(2, 2)

}
\seealso{
Other coordinate systems: 
\code{\link{cartesian_to_polar}()},
\code{\link{deg_to_rad}()},
\code{\link{euclidean_dist}()},
\code{\link{hausdorff_distance}()},
\code{\link{interpolate_measurements}()},
\code{\link{polar_to_cartesian}()},
\code{\link{quadrant_eye}()},
\code{\link{rad_to_deg}()},
\code{\link{rotate_shape}()},
\code{\link{signed_smallest_angle}()},
\code{\link{smallest_angle}()},
\code{\link{within_2pi}()},
\code{\link{within_360}()}
}
\concept{coordinate systems}
