% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{contour_to_sf_polygon}
\alias{contour_to_sf_polygon}
\title{Convert a contour to a polygon}
\usage{
contour_to_sf_polygon(contour, required_points = 4)
}
\arguments{
\item{contour}{A data frame containing one row for each point of the contour and the same
columns as \link{sample_contour}. If the contour needs to be scaled and/or
rotated, the necessary transforms must be performed prior to calling
\code{contour_to_sf_polygon()}.}

\item{required_points}{A number indicating the minimum number of points required to convert a
segment of the contour to a feature of the polygon. Must be four or more.}
}
\value{
An sf object. See details.
}
\description{
Converts a contour to an sf polygon object.
}
\details{
Each segment of the contour will become a feature of the polygon. There are
two exceptions, however: Segments that are fully contained within another
segment will \emph{subtracted} from the resulting polygon, and segments with
fewer than \code{required_points} will be disregarded.
}
\examples{
contour_to_sf_polygon(
  contour = get_contour(sample_curvature, contour_power = 44.5)
)

reference_contours |>
  dplyr::filter(cluster == 7) |>
  contour_to_sf_polygon()

}
\seealso{
Other Polygons: 
\code{\link{silhouette_overlap}()},
\code{\link{silhouette_overlap_group}()},
\code{\link{silhouette_plot}()}
}
\concept{Polygons}
