% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meridians.R
\name{curvature_meridians}
\alias{curvature_meridians}
\title{Find an exam's semi-meridians}
\usage{
curvature_meridians(
  exam_curvature,
  exam_astig,
  cornea_surface = "FRONT",
  interp = T
)
}
\arguments{
\item{exam_curvature}{A data frame having the same columns as \link{sample_curvature} and
containing one row for each curvature \code{measurement}.}

\item{exam_astig}{A data frame having the same columns as \link{sample_astig} and containing
one row for each of the three ring diameters.}

\item{cornea_surface}{A string, either 'FRONT' or 'BACK', indicating the anterior or posterior
surface of the cornea, respectively.}

\item{interp}{A Boolean. \code{TRUE} to interpolate measurements along the rings.}
}
\value{
A data frame containing one row for each semi-meridian.

\describe{
\item{ring}{Ring diameter of the semi-meridian.}
\item{radius}{Ending radius of the semi-meridian segment.}
\item{starting_radius}{Starting radius of the semi-meridian segment.}
\item{axis}{Either the \code{flat} or \code{steep} axis of astigmatism,
identifying the semi-meridian's reference axis.}
\item{between_min_max}{A Boolean identifying whether the semi-meridian was
found within a 180° window (±90°) of the opposite angle of astigmatism (or
the reflective "hemisphere").}
\item{theta}{Offset angle (in degrees) from the 3 o'clock position.}
\item{power}{Dioptric power of of the semi-meridian. When actual
measurement data are used, \code{power} will be the power at the end radius
of the semi-meridian segment. When interpreted data are used, \code{power}
will be the average power of the semi-meridian segment. }
}
}
\description{
Find the semi-meridians for an exam at three diameters (3, 5, and 7 mm).
}
\details{
This function identifies the radial arms with the highest and lowest average
power within 180° windows (±90°) of the axes of astigmatism (as determined
by the Pentacam) at the 3, 5, and 7 mm ring diameters.
}
\examples{
curvature_meridians(sample_curvature, sample_astig) |>
  dplyr::select(-tidyselect::all_of(join_fields))

}
\concept{Meridians}
