% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meridians.R
\name{curvature_meridians}
\alias{curvature_meridians}
\title{Find an exam's hemi-meridians}
\usage{
curvature_meridians(
  exam_curvature,
  exam_astig,
  cornea_surface = "FRONT",
  interp = T
)
}
\arguments{
\item{exam_curvature}{A data frame containing one row for each curvature \code{measurement} and the
same columns as \code{sample_curvature}.}

\item{exam_astig}{A data frame containing one row for each astigmatism \code{measurement} and
the same columns as \code{sample_astig}.}

\item{cornea_surface}{A string, either 'FRONT' or 'BACK', indicating the anterior or posterior
surface of the cornea, respectively.}

\item{interp}{A Boolean. \code{TRUE} to interpolate measurements along the rings.
\code{FALSE} is \emph{not} currently implemented.}
}
\value{
A data frame containing one row for each hemi-meridian.
}
\description{
Find the hemi-meridians for an exam at three diameters (3, 5, and 7 mm) and
four offset angles (0째, 90째, 180째, and 270째).
}
\examples{
curvature_meridians(sample_curvature, sample_astig)

}
\concept{Meridians}
