% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meridians.R
\name{curvature_meridians}
\alias{curvature_meridians}
\title{Find an exam's hemi-meridians}
\usage{
curvature_meridians(
  exam_curvature,
  exam_astig,
  cornea_surface = "FRONT",
  interp = T
)
}
\arguments{
\item{exam_curvature}{A data frame having the same columns as \link{sample_curvature} and
containing one row for each curvature \code{measurement}.}

\item{exam_astig}{A data frame having the same columns as \link{sample_astig} and containing
one row for each of the three ring diameters.}

\item{cornea_surface}{A string, either 'FRONT' or 'BACK', indicating the anterior or posterior
surface of the cornea, respectively.}

\item{interp}{A Boolean. \code{TRUE} to interpolate measurements along the rings.
\code{FALSE} is \emph{not} currently implemented.}
}
\value{
A data frame containing one row for each hemi-meridian.

\describe{
\item{meridian}{Either the \code{flat} or \code{steep} axis of astigmatism,
identifying the hemi-meridian's reference axis.}
\item{offset}{Offset angle (in degrees) from the 3 o'clock position.}
\item{ring_diam}{Radial end point of the hemi-meridian.}
\item{angle}{Angle (in degrees) of the hemi-meridian.}
\item{x}{\emph{x}-axis coordinate of the hemi-meridian's end point.}
\item{y}{\emph{y}-axis coordinate of the hemi-meridian's end point.}
\item{z}{Radius of curvature at the hemi-meridian's end point.}
\item{measurement}{Dioptric power of of the hemi-meridian at its end point.}
}
}
\description{
Find the hemi-meridians for an exam at three diameters (3, 5, and 7 mm) and
four offset angles (0°, 90°, 180°, and 270°).
}
\details{
This function looks at a 90° window (±45°) around each of the offset angles,
relative to the axis of astigmatism. For the two steep meridians, it finds
the min radius (i.e., max power) within the axes windows at each diameter
ring. For the two flat meridians, it finds the max radius (i.e., min power).
}
\examples{
curvature_meridians(sample_curvature, sample_astig) |>
  dplyr::select(-tidyselect::all_of(join_fields))

}
\concept{Meridians}
