% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{curvature_plot}
\alias{curvature_plot}
\title{Generate the curvature plot of the anterior surface of a cornea}
\usage{
curvature_plot(
  exam_curvature,
  basic = F,
  grid = F,
  labels = T,
  titles = T,
  axes = T,
  legend = T,
  truncate_circle = T,
  circle_dia = 9,
  points = F,
  greyscale = F
)
}
\arguments{
\item{exam_curvature}{A data frame containing one row for each curvature \code{measurement} and the
same columns as \link{sample_curvature}.}

\item{basic}{A Boolean. \code{TRUE} to forgo contours and colors.}

\item{grid}{A Boolean. \code{TRUE} to include a background grid.}

\item{labels}{A Boolean. \code{TRUE} to label the \emph{measured} curvature points.}

\item{titles}{A Boolean. \code{TRUE} to include plot title(s).}

\item{axes}{A Boolean. \code{TRUE} to include \emph{x} and \emph{y} axes.}

\item{legend}{A Boolean. \code{TRUE} to include a legend/key.}

\item{truncate_circle}{A Boolean. \code{TRUE} to limit the curvature map to a diameter of
\code{circle_dia}.}

\item{circle_dia}{A number. \code{TRUE} to include a legend/key.}

\item{points}{A Boolean. \code{TRUE} to points on the plot at the \emph{x}- and \emph{x}-
coordinates of the \emph{measured} points.}

\item{greyscale}{A Boolean. \code{TRUE} to render the contours in greyscale.}
}
\value{
A ggplot object.
}
\description{
Returns (but does not render) the curvature plot of the anterior surface of a
cornea by interpolating measurements from \code{COR-PWR.CSV}.
}
\details{
This function currently works for anterior surfaces only because
\code{plot_scale()} presumes the curvature maps uses the Pentacam's American
scale, which is designed for the dioptric power of anterior surfaces.
}
\examples{
curvature_plot(sample_curvature)

}
\seealso{
Other Plots: 
\code{\link{plot_scale}()}
}
\concept{Plots}
