% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_distributions.R
\name{d_mode}
\alias{d_mode}
\title{Dominant mode of a sample}
\usage{
d_mode(x)
}
\arguments{
\item{x}{A vector of numeric values.}
}
\value{
Returns the mode of \code{x}.
}
\description{
Calculates the mode for a vector of numeric values \code{x}.
}
\examples{
d_mode(rnorm(1000, mean = 2))

}
\seealso{
Other Statistical Functions: 
\code{\link{knee}()},
\code{\link{mean_folded_normal}()},
\code{\link{mean_half_normal}()},
\code{\link{pct_to_z}()},
\code{\link{sd_folded_normal}()},
\code{\link{sd_half_normal}()},
\code{\link{z_score}()},
\code{\link{z_to_pct}()}
}
\concept{Statistical Functions}
