% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_systems.R
\name{deg_to_rad}
\alias{deg_to_rad}
\title{Convert an angle from degrees to radians}
\usage{
deg_to_rad(angle_deg)
}
\arguments{
\item{angle_deg}{The angle in degrees.}
}
\value{
The angle in radians.
}
\description{
Convert an angle from degrees to radians.
}
\examples{
deg_to_rad(180)

}
\seealso{
Other Coordinate Systems: 
\code{\link{cartesian_degrees}()},
\code{\link{cartesian_to_polar}()},
\code{\link{euclidean_dist}()},
\code{\link{interpolate_measurements}()},
\code{\link{polar_to_cartesian}()},
\code{\link{rad_to_deg}()},
\code{\link{rotate_shape}()},
\code{\link{signed_smallest_angle}()},
\code{\link{smallest_angle}()},
\code{\link{within_360}()}
}
\concept{Coordinate Systems}
