% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_systems.R
\name{euclidean_dist}
\alias{euclidean_dist}
\title{Euclidean distance}
\usage{
euclidean_dist(x, y)
}
\arguments{
\item{x}{The \emph{x}-axis component.}

\item{y}{The \emph{y}-axis component.}
}
\value{
The distance of the point from the origin.
}
\description{
Calculates the Euclidean distance of a point (\emph{x}, \emph{y}) from the
origin (0, 0).
}
\examples{
euclidean_dist(2, 2)

}
\seealso{
Other Coordinate Systems: 
\code{\link{cartesian_degrees}()},
\code{\link{cartesian_to_polar}()},
\code{\link{deg_to_rad}()},
\code{\link{hausdorff_distance}()},
\code{\link{interpolate_measurements}()},
\code{\link{polar_to_cartesian}()},
\code{\link{quadrant_eye}()},
\code{\link{rad_to_deg}()},
\code{\link{rotate_shape}()},
\code{\link{signed_smallest_angle}()},
\code{\link{smallest_angle}()},
\code{\link{within_2pi}()},
\code{\link{within_360}()}
}
\concept{Coordinate Systems}
