% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contours.R
\name{fill_contour}
\alias{fill_contour}
\title{Fill a contour with points}
\usage{
fill_contour(exam_curvature, contour_power, interp = T, scale_rotate = T, ...)
}
\arguments{
\item{exam_curvature}{A data frame containing one row for each curvature \code{measurement} and the
same columns as \link{sample_curvature}.}

\item{contour_power}{A number defining the dioptric power of the contour to fill.}

\item{interp}{A Boolean. \code{TRUE} to interpolate points within the contour.}

\item{scale_rotate}{A Boolean. \code{TRUE} to scale and rotate the contour.}

\item{...}{Additional parameters to be passed to \code{scale_rotate()}, specifically
\code{axs} and \code{r_target}.}
}
\value{
A data frame with the same columns as \link{sample_curvature} containing
the points of the filled and (if \code{scale_rotate == T}) transformed contour.
}
\description{
Fills a contour with points, creating a silhouette of its shape.
}
\details{
Currently only handles the anterior (front) surface of the cornea.

Because the function creates data (through interpolation), it cannot return
any additional columns that may have been included as part of the incoming
\code{exam_curvature}.
}
\examples{
fill_contour(sample_curvature, contour_power = 45.5, axs = 36.2) |>
  head()

}
\seealso{
Other Contours: 
\code{\link{contour_polygon}()},
\code{\link{get_contour}()},
\code{\link{get_contour_levels}()},
\code{\link{scale_rotate}()},
\code{\link{silhouette_compare}()},
\code{\link{silhouette_compare_group}()}
}
\concept{Contours}
