% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contours.R
\name{get_contour}
\alias{get_contour}
\title{Get contour}
\usage{
get_contour(exam_curvature, add_edge = F, contour_power)
}
\arguments{
\item{exam_curvature}{A data frame containing one row for each curvature \code{measurement} and the
same columns as \link{sample_curvature}.}

\item{add_edge}{A Boolean. If the contour extends beyond the edge of the scanned cornea,
\code{TRUE} will add the edge of the scanned area, connecting the ends of the
contour and fully enclosing it. The \code{segment_id} for the added edge will
be \code{0}. If the contour doesn't extend beyond the edge of the cornea, no
edge will be added.}

\item{contour_power}{A number identifying the power of the contour to return. Valid values for an
exam can be determined by calling \link{get_contour_levels}.}
}
\value{
A data frame containing one row for each point of the contour and the same
columns as \link{sample_contour}.
}
\description{
Returns the \emph{x} and \emph{y} coordinates of the contour for an exam,
\code{exam_curvature}, at the specified power, \code{contour_power}.
}
\details{
A contour is comprised of one or more segments, uniquely identified by
\code{segment_id}.

This function doesn't currently work for posterior surfaces because
\link{plot_scale} presumes the color scale is on the "absolute
scale", which is designed for the dioptric power of anterior surfaces.
}
\examples{
get_contour(sample_curvature, contour_power = 45.5) |>
  head()

curvature_plot(sample_curvature, labels = F) +
  ggplot2::geom_point(data = get_contour(sample_curvature, contour_power = 45.5))

}
\seealso{
Other Contours: 
\code{\link{contour_polygon}()},
\code{\link{fill_contour}()},
\code{\link{get_contour_levels}()},
\code{\link{scale_rotate}()},
\code{\link{silhouette_compare}()},
\code{\link{silhouette_compare_group}()},
\code{\link{silhouette_plot_old}()}
}
\concept{Contours}
