% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contours.R
\name{get_contour_levels}
\alias{get_contour_levels}
\title{Get contour levels}
\usage{
get_contour_levels(exam_curvature)
}
\arguments{
\item{exam_curvature}{A data frame containing one row for each curvature \code{measurement} and the
same columns as \link{sample_curvature}.}
}
\value{
A table. The table's labels identify the contour levels, and the values in
the table identify the number of segments that comprise the contour.
}
\description{
Returns the full range of contours and the number of segments needed to plot
\code{exam_curvature} on a curvature map.
}
\details{
This function doesn't currently work for posterior surfaces because the
\code{plot_scale} function presumes the color scale is on the "absolute
scale", which is designed for the dioptric power of anterior surfaces.
}
\examples{
get_contour_levels(sample_curvature)

}
\seealso{
Other Contours: 
\code{\link{contour_polygon}()},
\code{\link{fill_contour}()},
\code{\link{get_contour}()},
\code{\link{scale_rotate}()},
\code{\link{silhouette_compare}()},
\code{\link{silhouette_compare_group}()}
}
\concept{Contours}
