% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_systems.R
\name{interpolate_measurements}
\alias{interpolate_measurements}
\title{Interpolate measurements over an expanded grid}
\usage{
interpolate_measurements(source_dat, ...)
}
\arguments{
\item{source_dat}{A data frame containing one row for each \code{measurement}. Each row must contain
at least three columns: a set of Cartesian coordinates (\code{x} and
\code{y}) and a \code{measurement}.}

\item{...}{Additional parameters that get passed to \code{akima::interp()}}
}
\value{
A new data frame with the interpolated measurements.
}
\description{
Interpolates measurements over an expanded
grid of Cartesian coordinates (\emph{x} and \emph{y}).
}
\details{
The \emph{x} and \emph{y} axes of the expanded grid will contain the same
number of unique values as original axes, but they will be equally spaced
along those axes.

The interpolated measurements will \emph{not} necessarily include the
original measurements.
}
\examples{
interpolate_measurements(sample_curvature) |>
   head()

}
\seealso{
Other Coordinate Systems: 
\code{\link{cartesian_degrees}()},
\code{\link{cartesian_to_polar}()},
\code{\link{deg_to_rad}()},
\code{\link{euclidean_dist}()},
\code{\link{polar_to_cartesian}()},
\code{\link{rad_to_deg}()},
\code{\link{rotate_shape}()},
\code{\link{signed_smallest_angle}()},
\code{\link{smallest_angle}()},
\code{\link{within_360}()}
}
\concept{Coordinate Systems}
