% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_systems.R
\name{interpolate_measurements}
\alias{interpolate_measurements}
\title{Interpolate measurements over an expanded grid}
\usage{
interpolate_measurements(source_dat, radius = 4.5, ...)
}
\arguments{
\item{source_dat}{A data frame containing one row for each \code{measurement} (i.e., radius of
curvature). Each row must contain at least three columns: a set of Cartesian
coordinates (\code{x} and \code{y}) and a \code{measurement}.}

\item{radius}{A number. Extent of the corneal radius to interpolate.}

\item{...}{Additional parameters that get passed to \code{akima::interp()}}
}
\value{
A new data frame with the interpolated measurements.
}
\description{
Interpolates measurements over an expanded
grid of Cartesian coordinates (\emph{x} and \emph{y}).
}
\details{
This function interpolates measurements at the same 0.1 mm grain as the
Pentacam's more detailed files. Rather than going from -5 to 5 along the
\code{x} and \code{y} axes (like the Petacam), the resulting grid will go
from -\code{radius} to +\code{radius}.

The interpolated measurements will \emph{not} necessarily include the
original measurements.
}
\examples{
interpolate_measurements(sample_curvature) |>
   head()

}
\seealso{
Other coordinate systems: 
\code{\link{cartesian_degrees}()},
\code{\link{cartesian_to_polar}()},
\code{\link{deg_to_rad}()},
\code{\link{euclidean_dist}()},
\code{\link{hausdorff_distance}()},
\code{\link{polar_to_cartesian}()},
\code{\link{quadrant_eye}()},
\code{\link{rad_to_deg}()},
\code{\link{rotate_shape}()},
\code{\link{signed_smallest_angle}()},
\code{\link{smallest_angle}()},
\code{\link{within_2pi}()},
\code{\link{within_360}()}
}
\concept{coordinate systems}
