% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_next.R
\name{k_next}
\alias{k_next}
\title{Finds k-next}
\usage{
k_next(exam_curvature, just_points = F)
}
\arguments{
\item{exam_curvature}{A data frame containing one row for each curvature \code{measurement} and the
same columns as \code{sample_curvature}.}

\item{just_points}{A Boolean. \code{TRUE} to return the points that comprise the primary peak.}
}
\value{
If \code{just_points} = \code{F}, \code{k_next()} returns a one-row data
frame containing the \code{join_fields} and a variety of metadata related to
k-next.

If \code{just_points} = \code{T}, \code{k_next()} returns a copy of
\code{exam_curvature} with an additional column \code{peak_id} identifying
the peak that each measured point belongs to.
}
\description{
Find the secondary peak on a curvature map.
}
\examples{
k_next(sample_curvature)

}
