% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_distributions.R
\name{mean_folded_normal}
\alias{mean_folded_normal}
\title{Mean of a folded-normal distribution}
\usage{
mean_folded_normal(mu, sigma)
}
\arguments{
\item{mu}{Mean of the normal distribution.}

\item{sigma}{Standard deviation of the normal distribution.}
}
\value{
Returns the mean.
}
\description{
Calculates the mean (expected value) of a folded-normal distribution based on
the parameters of a normal distribution. Presumes that the folded-normal
distribution is the absolute value of the normal distribution.
}
\examples{
mean_folded_normal(mu = 3, sigma = 1)

}
\seealso{
Other Statistical Functions: 
\code{\link{d_mode}()},
\code{\link{knee}()},
\code{\link{mean_half_normal}()},
\code{\link{pct_to_z}()},
\code{\link{sd_folded_normal}()},
\code{\link{sd_half_normal}()},
\code{\link{z_score}()},
\code{\link{z_to_pct}()}
}
\concept{Statistical Functions}
