% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_distributions.R
\name{mean_half_normal}
\alias{mean_half_normal}
\title{Mean of a half-normal distribution}
\usage{
mean_half_normal(x, ...)
}
\arguments{
\item{x}{A numeric vector of values from a half-normal distribution.}
}
\value{
Returns the mean of \code{x}.
}
\description{
Calculates the mean (expected value) of a sample from a half-normal
distribution.
}
\details{
\code{mean_half_normal()} uses \code{sd(x)} to account for the characteristics
of the half-normal distribution
}
\examples{
mean_half_normal(abs(rnorm(10)))

}
\seealso{
Other statistical functions: 
\code{\link{d_mode}()},
\code{\link{knee}()},
\code{\link{mean_folded_normal}()},
\code{\link{pct_to_z}()},
\code{\link{sd_folded_normal}()},
\code{\link{sd_half_normal}()},
\code{\link{z_score}()},
\code{\link{z_to_pct}()}
}
\concept{statistical functions}
