% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_scale}
\alias{plot_scale}
\title{Returns breaks on the \emph{z}-axis scale used to plot curvature maps}
\usage{
plot_scale(z)
}
\arguments{
\item{z}{A numeric vector containing curvature measurements (in diopters). Currently
works for \emph{anterior} surfaces only.}
}
\value{
A numeric vector containing the breaks on the \emph{z}-axis scale (in
diopters).
}
\description{
Returns breaks for \code{z} (presumed to be in diopters) on the dioptric
scale used to plot curvature maps.
}
\details{
The contour levels on a curvature map are defined by the breaks of the
color scale used to generate the map.

\code{plot_scale()} presumes the curvature maps uses the Pentacam's American
scale, which is designed for the dioptric power of anterior surfaces.
}
\examples{
anterior_power(sample_curvature$measurement) |>
  plot_scale()

}
\seealso{
Other Plots: 
\code{\link{curvature_plot}()}
}
\concept{Plots}
