% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_systems.R
\name{rad_to_deg}
\alias{rad_to_deg}
\title{Convert an angle from radians to degrees}
\usage{
rad_to_deg(angle_rad)
}
\arguments{
\item{angle_rad}{The angle in radians.}
}
\value{
The angle in degrees.
}
\description{
Convert an angle from radians to degrees.
}
\examples{
rad_to_deg(pi)

}
\seealso{
Other Coordinate Systems: 
\code{\link{cartesian_degrees}()},
\code{\link{cartesian_to_polar}()},
\code{\link{deg_to_rad}()},
\code{\link{euclidean_dist}()},
\code{\link{interpolate_measurements}()},
\code{\link{polar_to_cartesian}()},
\code{\link{rotate_shape}()},
\code{\link{signed_smallest_angle}()},
\code{\link{smallest_angle}()},
\code{\link{within_360}()}
}
\concept{Coordinate Systems}
