% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_systems.R
\name{signed_smallest_angle}
\alias{signed_smallest_angle}
\title{The smallest signed angle (in degrees) between two radial lines}
\usage{
signed_smallest_angle(angle_1, angle_2)
}
\arguments{
\item{angle_1}{The angle (in degrees) of the first radial line.}

\item{angle_2}{The angle (in degrees) of the second radial line.}
}
\value{
The angle (in degrees) between with the two lines.
}
\description{
Calculates the smallest angle (in degrees) between two radial lines. If
\code{angle_1} > \code{angle_2}, the result is negative. If
\code{angle_1} < \code{angle_2}, the result is positive.
}
\details{
The result is essentially the signed difference between two angles.
}
\examples{
signed_smallest_angle(45, 90)

}
\seealso{
\code{\link[=smallest_angle]{smallest_angle()}}
}
