% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contours.R
\name{silhouette_compare}
\alias{silhouette_compare}
\title{Compare the silhouettes of two contours}
\usage{
silhouette_compare(shape_A, polygon_A, shape_B, polygon_B, show_plots = F)
}
\arguments{
\item{shape_A}{A data frame containing the silhouette (filled shape) for contour A.}

\item{polygon_A}{A data frame containing the polygon object for contour A.}

\item{shape_B}{A data frame containing the silhouette (filled shape) for contour B.}

\item{polygon_B}{A data frame containing the polygon object for contour B.}

\item{show_plots}{A Boolean. \code{TRUE} to render two plots that illustrate the comparisons
being made.}
}
\value{
The percentage of overlap between the two silhouettes, averaged over both
directions.
}
\description{
Each of silhouettes is defined by two objects, a silhouette (filled shape)
and a polygon. These objects must be created before calling
\code{silhouette_compare()}. Silhouettes are created by calling
\link{fill_contour}, and polygons are created by calling
\link{contour_polygon}.
}
\examples{
silhouette_compare(
  shape_A = fill_contour(sample_curvature, contour_power = 45.5),
  polygon_A = contour_polygon(sample_contour),
  shape_B = fill_contour(sample_curvature, contour_power = 45.5),
  polygon_B = contour_polygon(sample_contour)
)

}
\seealso{
Other Contours: 
\code{\link{contour_polygon}()},
\code{\link{fill_contour}()},
\code{\link{get_contour}()},
\code{\link{get_contour_levels}()},
\code{\link{scale_rotate}()},
\code{\link{silhouette_compare_group}()}
}
\concept{Contours}
