% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contours.R
\name{silhouette_compare_group}
\alias{silhouette_compare_group}
\title{Compare the silhouette of one contour with a group of silhouettes}
\usage{
silhouette_compare_group(
  ref_shape,
  ref_polygon,
  group_shapes,
  group_polygons,
  show_plots = F,
  return_detail = F,
  ...
)
}
\arguments{
\item{ref_shape}{A data frame containing the filled shape for the reference contour}

\item{ref_polygon}{A data frame containing the polygon object for the reference contour}

\item{group_shapes}{A data frame containing filled shapes for the group of contours.
\emph{Requires} a \code{cluster} column that uniquely identifies
each member of the group. The cluster IDs for individual contours (i.e.,
members of the group) must be the same in both \code{group_shapes} and
\code{group_polygons}.}

\item{group_polygons}{A data frame containing polygon objects for the group of contours.
\emph{Requires} a \code{cluster} column that uniquely identifies
each member of the group. The cluster IDs for individual contours (i.e.,
members of the group) must be the same in both \code{group_shapes} and
\code{group_polygons}.}

\item{show_plots}{A Boolean. \code{TRUE} to render two plots that illustrate the comparisons
being made.}

\item{return_detail}{A Boolean. \code{TRUE} to return a list object with two elements:
\code{closest_fit}, the index of the group member that has the highest
percentage of overlap with the reference silhouette, and
\code{average_distance}, a vector containing the average percentage of
overlap between the reference silhouette and every member of the group.}
}
\value{
The index of the group member that has the highest percentage of overlap with
the reference silhouette.
}
\description{
Every silhouette is defined by two objects, a filled shape and a polygon.
These objects must be created before calling
\code{silhouette_compare_group()}. The filled shapes are created by calling
\link{fill_contour}, and the polygon objects are created by calling
\link{contour_polygon}.

Each of the comparisons is made in both directions.
}
\details{
Both \code{group_shapes} and \code{group_polygons} require a \code{cluster}
column. Although the cluster column can be passed into and returned by
\link{contour_polygon}, the same cannot be done by \link{fill_contour},
which means the process for creating \code{group_shapes} requires the
additional step of appending the \code{cluster} column.
}
\seealso{
Other Contours: 
\code{\link{contour_polygon}()},
\code{\link{fill_contour}()},
\code{\link{get_contour}()},
\code{\link{get_contour_levels}()},
\code{\link{scale_rotate}()},
\code{\link{silhouette_compare}()}
}
\concept{Contours}
