% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contours.R
\name{silhouette_compare_group}
\alias{silhouette_compare_group}
\title{Compare the silhouette of one contour with the silhouettes of a group}
\usage{
silhouette_compare_group(
  ref_shape,
  ref_polygon,
  group_shapes,
  group_polygons,
  show_plots = F,
  return_detail = F,
  ...
)
}
\arguments{
\item{ref_shape}{A data frame containing the silhouette (filled shape) for the exam.}

\item{ref_polygon}{A data frame containing the polygon object for the exam.}

\item{group_shapes}{A data frame containing silhouettes (filled shapes) for the group of
contours. The data frame \emph{must} include a \code{cluster} column that
uniquely identifies each member of the group. The cluster IDs for individual
contours (i.e., members of the group) must be the same in both
\code{group_shapes} and \code{group_polygons}.}

\item{group_polygons}{A data frame containing polygons for the group of contours. The data frame
\emph{must} include a \code{cluster} column that uniquely identifies
each member of the group. The cluster IDs for individual contours (i.e.,
members of the group) must be the same in both \code{group_shapes} and
\code{group_polygons}.}

\item{show_plots}{A Boolean. \code{TRUE} to render a plot that illustrates the comparisons
being made. To access the plot as an object, see \link{IQeyes:silhouette_plot_old}.}

\item{return_detail}{A Boolean. \code{FALSE} (the default) to return the index of the group member
that has the highest percentage of overlap with the reference silhouette.
\code{TRUE} to return a list object with two elements: \code{closest_fit} and
\code{overlap}, a vector containing the average percentage of
overlap between the reference silhouette and every member of the group.}
}
\value{
If \code{return_detail = F}, the function returns just the index of the
group member that has the highest percentage of overlap with
the reference silhouette.

If \code{return_detail = T}, the data frame will be a list object object with
two elements: \code{closest_fit}, the index of the group member with the
highest percentage of overlap, and \code{overlap}, a vector containing the
percentage overlap between the reference silhouette and every member of the
group.
}
\description{
Every silhouette is defined by two objects, the silhouette (filled shape) and
a polygon (an outline with connected ends). These objects must be created
before calling \code{silhouette_compare_group()}. Silhouettes (filled shapes)
are created by calling \link{fill_contour}, and polygons are created by
calling \link{contour_polygon}.

Each of the comparisons is made in both directions.
}
\details{
The data frames passed into \code{group_shapes} and \code{group_polygons}
both require a \code{cluster} column. Although a cluster column can be
passed into and returned by \link{contour_polygon}, the same cannot be
done by \link{fill_contour}, which means the process for creating
\code{group_shapes} requires the additional step of appending the
\code{cluster} column.

For convenience, the silhouettes \link{IQeyes:canonical_silhouettes} and polygons
\link{IQeyes:canonical_polygons} for the canonical shapes are included as
objects in this package.
}
\examples{
sample_polygon <- get_contour(sample_curvature, add_edge = T, contour_power = 45.5) |>
  scale_rotate(axs = 33.6) |>
  contour_polygon()

silhouette_compare_group(
  ref_shape = fill_contour(sample_curvature, contour_power = 45.5, axs = 33.6),
  ref_polygon = sample_polygon,
  group_shapes = canonical_silhouettes,
  group_polygons = canonical_polygons,
  show_plots = T,
  return_detail = T
)

}
\seealso{
Other Contours: 
\code{\link{contour_polygon}()},
\code{\link{fill_contour}()},
\code{\link{get_contour}()},
\code{\link{get_contour_levels}()},
\code{\link{scale_rotate}()},
\code{\link{silhouette_compare}()},
\code{\link{silhouette_plot_old}()}
}
\concept{Contours}
