% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{silhouette_plot}
\alias{silhouette_plot}
\title{Plot the overlap of one silhouette with one or more silhouettes}
\usage{
silhouette_plot(ref_polygon, compare_polygons, highlight_cluster = NULL)
}
\arguments{
\item{ref_polygon}{A data frame containing the polygon object for the exam.}

\item{compare_polygons}{A data frame containing one or more polygons. If the data frame contains a
\code{cluster} column, the plot will be faceted by cluster.}
}
\value{
A ggplot2 object that that shows the overlapping silhouettes.
}
\description{
Plots the overlap of one contour's silhouette with either another silhouette
or a group of silhouettes.

Plotting requires sf polygon objects. These objects can be created by
calling \link{contour_to_sf_polygon} before calling
\code{silhouette_plot()}.
}
\details{
For convenience, the polygons for the reference shapes
(\link{IQeyes:reference_polygons}) are included in this package.
}
\examples{
sample_polygon <- get_contour(sample_curvature, add_edge = T, contour_power = 45.5) |>
  scale_rotate(axs = 33.6) |>
  contour_to_sf_polygon()

silhouette_plot(
  ref_polygon = sample_polygon,
  compare_polygons = reference_polygons,
  highlight_cluster = 1
)

}
\seealso{
Other Polygons: 
\code{\link{contour_to_sf_polygon}()},
\code{\link{silhouette_overlap}()},
\code{\link{silhouette_overlap_group}()}
}
\concept{Polygons}
