% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_systems.R
\name{within_360}
\alias{within_360}
\title{Keep an angle within the 0--360° range}
\usage{
within_360(angle)
}
\arguments{
\item{angle}{An angle in degrees.}
}
\value{
The equivalent angle within the 0--360° range.
}
\description{
Converts any angle (in degrees) to the equivalent angle in the 0--360° range.
}
\details{
Adding or subtracting two angles can sometimes result in a new angle that is
outside the 0--360° range. This function converts the result to the equivalent
angle within the 0--360° range.
}
\examples{
within_360(551)
within_360(42)
within_360(-22)

}
\seealso{
Other Coordinate Systems: 
\code{\link{cartesian_degrees}()},
\code{\link{cartesian_to_polar}()},
\code{\link{deg_to_rad}()},
\code{\link{euclidean_dist}()},
\code{\link{hausdorff_distance}()},
\code{\link{interpolate_measurements}()},
\code{\link{polar_to_cartesian}()},
\code{\link{quadrant_eye}()},
\code{\link{rad_to_deg}()},
\code{\link{rotate_shape}()},
\code{\link{signed_smallest_angle}()},
\code{\link{smallest_angle}()},
\code{\link{within_2pi}()}
}
\concept{Coordinate Systems}
